Apply mypy config override to blocked_signals
