+++ tests/lib/server.py
@@ -1,5 +1,4 @@
 import os
-import signal
 import ssl
 import threading
 from base64 import b64encode
@@ -11,7 +10,7 @@ from unittest.mock import Mock
 from werkzeug.serving import BaseWSGIServer, WSGIRequestHandler
 from werkzeug.serving import make_server as _make_server
 
-from .compat import nullcontext
+from .compat import blocked_signals
 
 if TYPE_CHECKING:
     from _typeshed.wsgi import StartResponse, WSGIApplication, WSGIEnvironment
@@ -21,37 +20,6 @@ Body = Iterable[bytes]
 
 class MockServer(BaseWSGIServer):
     mock: Mock = Mock()
-
-
-# Applies on Windows.
-if not hasattr(signal, "pthread_sigmask"):
-    # We're not relying on this behavior anywhere currently, it's just best
-    # practice.
-    blocked_signals = nullcontext
-else:
-
-    @contextmanager
-    def blocked_signals() -> Iterator[None]:
-        """Block all signals for e.g. starting a worker thread."""
-        # valid_signals() was added in Python 3.8 (and not using it results
-        # in a warning on pthread_sigmask() call)
-        mask: Iterable[int]
-        try:
-            mask = signal.valid_signals()
-        except AttributeError:
-            mask = set(range(1, signal.NSIG))
-
-        old_mask = signal.pthread_sigmask(  # type: ignore[attr-defined]
-            signal.SIG_SETMASK,  # type: ignore[attr-defined]
-            mask,
-        )
-        try:
-            yield
-        finally:
-            signal.pthread_sigmask(  # type: ignore[attr-defined]
-                signal.SIG_SETMASK,  # type: ignore[attr-defined]
-                old_mask,
-            )
 
 
 class _RequestHandler(WSGIRequestHandler):
